map now display children zones
